infra/requirements.txt
aws-cdk-lib>=2.170.0
constructs>=10.0.0